+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed #43: Renamed `WsjcppCore::dirExists` to `wsjcpp::dir_exists`
 - Fixed #44: Renamed `WsjcppCore::fileExists` to `wsjcpp::file_exists`
+- Fixed #46: Rename `WsjcppCore::split` to `wsjcpp::split`
 
 ## [v0.2.6] - 2026-07-05 (2026 July 5)